requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
pdf2image
python-docx
